+++ packages/codemirror-block-editor/src/index.test.ts
@@ -1,5 +1,0 @@
-describe("Index", () => {
-  it("should pass", () => {
-    expect(true).toEqual(true);
-  });
-});