+++ packages/codemirror-block-editor/src/index.ts
@@ -6,15 +6,12 @@ import { writeVersionToHtml } from "./write-version-to-html";
 writeVersionToHtml();
 
 const initialState = EditorState.create({
-  doc: [
-    "- Indent block with Tab",
-    "- and use Shift-Tab to decrease the block indentation",
-    "- itemA",
-    "  - itemB",
-    "    - itemC",
-    "- itemD",
-  ].join("\n"),
-  extensions: [vim(), blockLevelExtension(), basicSetup],
+  doc: ["- a"].join("\n"),
+  extensions: [
+    // vim(),
+    blockLevelExtension(),
+    basicSetup,
+  ],
 });
 const view = new EditorView({
   parent: document.getElementById("editor") ?? undefined,